+++ test/editor/markdown_document_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_smooth_markdown/flutter_smooth_markdown.dart';
+import 'package:flutter_smooth_markdown/flutter_smooth_markdown_editor_experimental.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void _expectUniqueDocumentIds(MarkdownDocument document) {